requirements.txt
# requirements.txt
boto3==1.37.34
botocore==1.37.34
duckdb==1.2.2
numpy==2.2.4
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.11.4
urllib3==2.4.0